Create PORTUGUESE.json
Hi, guys. I am Brazilian and, in order to collaborate with the DisCal team, I translated it into Portuguese. I hope it will be accepted as it has taken me considerable time to adapt and revise. I leave here my thanks to the developers for the creation and maintenance of DisCal.